% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above_percent.R
\name{above_percent}
\alias{above_percent}
\title{Calculate percentage above targeted values}
\usage{
above_percent(data, targets = c(140, 180, 200, 250))
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{targets}{List of target values. Default list is (140,180,200,250).}
}
\value{

}
\description{
Calculate percentage above targeted values
}
\examples{
above_percent(data)
above_percent(data, targets = c(100, 150, 180))

}
