% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/below_percent.R
\name{below_percent}
\alias{below_percent}
\title{Calculate percentage below targeted values}
\usage{
below_percent(data, targets = c(50, 80))
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{targets}{List of target values. Default list is (50,80).}
}
\value{

}
\description{
Calculate percentage below targeted values
}
\examples{
below_percent(data)
below_percent(data, targets = c(50,100, 180))
}
