% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_hall}
\alias{example_data_hall}
\title{Example data from Hall et al. (2018)}
\format{
a data.frame with 5567 rows and 3 columns, which are:
\describe{
\item{id}{identifier of subject}
\item{time}{date and time stamp}
\item{gl}{glucose level as measured by CGM (mg/dL)}
}
}
\usage{
example_data_hall
}
\description{
Three subjects from the Hall publicly available dataset,
primarily intended for use with example_meals_hall
}
\keyword{datasets}
