% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_meal}
\alias{example_meal}
\title{Example CGM mealtime data}
\format{
A data.frame with 1 row and 3 columns, which are:
\describe{
\item{id}{identifier of subject}
\item{meal}{meal type identifier}
\item{mealtime}{time of meal}
}
}
\usage{
example_meal
}
\description{
Example of mealtimes data format for meal_metrics function,
corresponds to example_data_1_subject data.
}
\keyword{datasets}
