% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{grade}
\alias{grade}
\title{Calculate mean GRADE score}
\usage{
grade(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").
1 represents mmol/L. Default is 0 (mg/dL).}
}
\description{
Calculate mean GRADE score
}
\examples{
grade(data)
}
