% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_eugly.R
\name{grade_eugly}
\alias{grade_eugly}
\title{Percentage of GRADE score in target range}
\usage{
grade_eugly(data, lower = 80, upper = 140)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{lower}{Lower bound used for hypoglycemia cutoff, in mg/dL. Default is 80}

\item{upper}{Upper bound used for hyperglycemia cutoff, in mg/dL. Default is 140.}
}
\value{

}
\description{
Percentage of GRADE score in target range
}
\examples{
grade_eugly(data)
grade_eugly(data, lower = 70, upper = 180)
}
