% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_hypo.R
\name{grade_hypo}
\alias{grade_hypo}
\title{Percentage of GRADE score attributable to hypoglycemia}
\usage{
grade_hypo(data, lower = 70)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. NA's will be omitted from the glucose
values in calculation of \%GRADE hypoglycemia.}

\item{lower}{Lower bound used for hypoglycemia cutoff, in mg/dL. Default is 70}
}
\value{

}
\description{
The function grade_hypo produces \%GRADE hypoglycemia values in data.frame
form with one column and one row per subject.
}
\details{
A dataframe structure with one column and a row for each subject.

\%GRADE hypoglycemia is calculated by calculating the GRADE score (see grade
function) just for values below the hypoglycemia cutoff and dividing by the
total GRADE score.

Wrapping as.numeric() around the grade_hypo call on a dataset with
a single subject will return a numeric value corresponding to the \%GRADE
hypoglycemia value. This will not work for datasets with multiple subjects.
}
\examples{

data(example_data_1_subject)
grade_hypo(example_data_1_subject)
grade_hypo(example_data_1_subject, lower = 80)

data(example_data_5_subject)
grade_hypo(example_data_5_subject)
grade_hypo(example_data_5_subject, lower = 65)

}
\references{
Rodbard (2009) Interpretation of continuous glucose monitoring data:
glycemic variability and quality of glycemic control,
\emph{Diabetes Technology and Therapeutics} \strong{11 Suppl 1},
S55-67. \doi{10.1089/dia.2008.0132}.
}
