% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_sensors.r
\name{importdexcom}
\alias{importdexcom}
\title{Functions to read raw data from a variety of common sensors. Assumes .csv format for all data. Due to limited accessiblity of data,
only importlibre has been tested on actual data as of 4/7/2021.}
\usage{
importlibre("cgmdata.csv", id = "read")
}
\arguments{
\item{filename}{String matching the name of the data to be read. Assumed to be .csv}

\item{id}{String indicating subject id. Defaults to "read" or "filename" depending on the import function.
A value of "read" will cause the program to attempt to read the subject id from the file. A value of "filename" will cause the
program to use the basename of the filename (i.e. filename without any directory information) with .csv removed as subject id.}
}
\value{
A dataframe containing the data read from the provided file.
}
\description{
Several helper functions to assist in reading data directly from sensor outputs. Should return a dataframe in correct format
for use with the rest of the \code{iglu} package. Assumes all data will be readable with base R read.csv function.
}
\details{
Heavily derived from the readers available in the cgmanalysis package's cleandata function.

A dataframe object with the columns "id", "time" and "gl" and one row per reading will be returned.
}
\examples{
importdexcom("johndoe.csv", id = "filename")

importlibre("yourdata.csv", id = "read")

importlibrepro("data.csv" id = "read")

importasc("home/janedoe.csv", id = "filename")

importipro("cgmsequences.csv", id = "read")

}
\references{
Vigers et al. (2019) cgmanalysis: An R package for descriptive analysis of continuous glucose monitor data
\emph{PLoS ONE} \strong{14(10)}: e0216851,
\doi{10.1371/journal.pone.0216851}
}
