% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_glu.R
\name{iqr_glu}
\alias{iqr_glu}
\title{Calculate glucose IQR}
\usage{
iqr_glu(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}
}
\value{

}
\description{
Simple wrapper for IQR().
}
\examples{
iqr_glu(data)
}
