% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_index.R
\name{j_index}
\alias{j_index}
\title{Calculate J-index}
\usage{
j_index(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").
1 represents mmol/L. Default is 0 (mg/dL).}
}
\description{
The J-index is calculated differently for each of the
two measurement units. For mg/dL (unit = 0), J-index = 0.001 * (mean + sd)^2
For mmol/L (unit = 1), J-index = 0.324 * (mean + sd)^2
}
\examples{
j_index(data)
}
