% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbgi.R
\name{lbgi}
\alias{lbgi}
\title{Calculate Low Blood Glucose Index (LGBI)}
\usage{
lbgi(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}
}
\value{

}
\description{
Calculate Low Blood Glucose Index (LGBI)
}
\examples{
lbgi(data)
}
