% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_glu.R
\name{mean_glu}
\alias{mean_glu}
\title{Calculate mean glucose level}
\usage{
mean_glu(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}
}
\description{
Simple wrapper for mean().
}
\examples{
mean_glu(data)
}
