% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_glu.R
\name{median_glu}
\alias{median_glu}
\title{Calculate median glucose level}
\usage{
median_glu(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}
}
\value{

}
\description{
Simple wrapper for median().
}
\examples{
median_glu(data)
}
