% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_scatterplot.R
\name{metric_scatter}
\alias{metric_scatter}
\title{Plot scatterplot of metric comparisons}
\usage{
metric_scatter(data, metric1 = "adrr", metric2 = "conga",
  subjects = NULL, lower = 70, upper = 140, targets_above = c(140,
  180, ))
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}

\item{metric2}{}
}
\value{

}
\description{
Plot scatterplot of metric comparisons
}
