% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_summary.R
\name{metric_summary}
\alias{metric_summary}
\title{Create table of metrics}
\usage{
metric_summary(data, metrics = NULL)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{metrics}{List of metric names to use from the following list:}
}
\value{

}
\description{
Create table with all or a (user selected) subset of metrics calculated by iglu functions. Currently
supports metrics not requiring interpolation, and with default options for each function call.
}
\examples{
metric_summary(data)
metric_summary(data, metrics = c("above_percent", "below_percent"))
}
