% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modd.R
\name{modd}
\alias{modd}
\title{Calculate mean difference between glucose values obtained at the same time
of day (MODD)}
\usage{
modd(data, lag = 1)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}

\item{lag}{Integer indicating which lag (# days) to use. Default is 1.}
}
\value{

}
\description{
Calculate mean difference between glucose values obtained at the same time
of day (MODD)
}
\examples{
modd(data)
modd(data, lag = 2)
}
