% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_quantile_glu.R
\name{multi_quantile_glu}
\alias{multi_quantile_glu}
\title{Calculate glucose level quantiles for multiple subjects}
\usage{
multi_quantile_glu(data, quantiles = c(0, 0.25, 0.5, 0.75, 1),
  subjects = unique(data$id))
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),}

\item{quantiles}{List of values between 0 and 1.}

\item{subjects}{List of subject id's}
}
\value{

}
\description{
Simple wrapper for quantile().
}
\examples{
multi_quantile_glu(data)
multi_quantile_glu(data, quantiles = c(0, .33, .66, 1))
multi_quantile_glu(data, subjects = 'Subject 1')
}
