% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Data Pre-Processor}
\usage{
process_data(data, id = "id", timestamp = "time", glu = "gl", time_parser = as.POSIXct)
}
\arguments{
\item{data}{User-supplied dataset containing continuous glucose monitor data. Must
contain data for time and glucose readings at a
minimum.}

\item{timestamp}{Required column name (character string) corresponding to time values in data. The dates can be
in any format parsable by as.POSIXct, or any format accepted by the parser passed to time_parser.}

\item{glu}{Required column name (character string) corresponding to blood glucose values, mg/dl}

\item{time_parser}{Optional function used to convert datetimes to time objects. Defaults to as.POSIXct.}
}
\value{
A processed dataframe object that cooperates with every other
function within the iglu package - all column names and formats comply.
}
\description{
A helper function to assist in pre-processing the user-supplied
input data for use with other functions.
Typically, this function will process the data and be stored as another
dataframe. This function ensures that the supplied data complies with every
function within the \code{iglu}   package. See Vignette for further details.
}
\examples{
data("example_data_1_subject")

# Process example data
processed <- process_data(example_data_1_subject, id = "id", timestamp = "time", glu = "gl", time_parser = as.PoSIXct)

processed

data("example_data_5_subject")

# Process example data
processed_multisubject <- process_data(example_data_5_subject, id = "id", timestamp = "time", glu = "gl", time_parser = as.PoSIXct)

processed_multisubject

}
