% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_glu.R
\name{range_glu}
\alias{range_glu}
\title{Calculate glucose level range}
\usage{
range_glu(data)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. NA's will be omitted from the glucose
values in calculation of the range.}
}
\value{

}
\description{
The function range_glu outputs the distance between minimum
and maximum glucose values per subject. The output is in a data.frame form
by default, with one column and a row corresponding to each subject.
}
\details{
A dataframe structure with 1 row for each subject and 1 column
for the range value is returned.

Wrapping as.numeric() around the range_glu call on a dataset with
a single subject will return a numeric value corresponding to the range.
This will not work for datasets with multiple subjects.
}
\examples{
data(example_data_1_subject)
range_glu(example_data_1_subject)

data(example_data_5_subject)
range_glu(example_data_5_subject)

}
