% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Calculate the Rate of Change at each time point (ROC)}
\usage{
roc(data, timelag = 15, tz = "")
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl"}

\item{timelag}{Integer indicating the time period (# minutes) over which rate of change
is calculated. Default is 15, e.g. rate of change is the change in glucose
over the past 15 minutes divided by 15.}

\item{tz}{A character string specifying the time zone to be used.
System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone,
and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC,
on some platforms with a warning.}
}
\value{
A tibble object with two columns: subject id and rate of change values
for each time point that it is available.
}
\description{
The function roc produces rate of
change values in a tibble object.
}
\details{
A tibble object with a column for subject id and a column for ROC values is returned. A ROC value
is returned for each recorded point for all the subjects. If the rate of change cannot be calculated,
the function will return NA for that point.

Missing values will be linearly interpolated when close enough to non-missing values.

The ROC is calculated as \eqn{\frac{BG(t_i) - BG(t_{i-1})}{t_i - t_{i-1}}}
where BG_i is the Blood Glucose measurement at time t_i and BG_{i-1} is the Blood Glucose
measurement at time t_{i-1}. The time difference between the points, t_i - t_{i-1}, is selectable
and set at a default of 15 minutes.
}
\examples{

data(example_data_1_subject)
roc(example_data_1_subject)
roc(example_data_1_subject, timelag = 10)

data(example_data_5_subject)
roc(example_data_5_subject)
roc(example_data_5_subject, timelag = 10)

}
\references{
Clarke et al. (2009) Statistical Tools to Analyze Continuous Glucose Monitor Data,
Diabetes
\emph{Diabetes Technology and Therapeutics} \strong{11} S45-S54,
\doi{10.1089/dia.2008.0138}.
}
