% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above_percent.R
\name{above_percent}
\alias{above_percent}
\title{Calculate percentage of values above target thresholds}
\usage{
above_percent(data, targets_above = c(140, 180, 200, 250))
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. NA's will be omitted from the glucose
values in calculation of percent.}

\item{targets_above}{Numeric vector of glucose thresholds. Glucose values from
data argument will be compared to each value in the targets_above vector.
Default list is (140, 180, 200, 250).}
}
\value{
A data.frame with columns of the percentage of values above a
threshold
}
\description{
The function above_percent produces a dataframe of values equal to
the percentage of glucose measurements above target values. The output is in
data.frame form by default, where the columns correspond to the target
values and the output rows correspond to the subjects. The values will be
between 0 (no measurements) and 100 (all measurements).
}
\details{
A dataframe structure with 1 row for each subject and a
column for each target value is returned.

Wrapping as.numeric() around the above_percent call on a dataset with
a single subject will return a numeric vector, where the values
correspond to the percent of glucose values above each threshold
in the order passed in the targets_above argument. This will not work for
datasets with multiple subjects.
}
\examples{

data(example_data_1_subject)

above_percent(example_data_1_subject)
above_percent(example_data_1_subject$gl)
above_percent(example_data_1_subject, targets_above = c(100, 150, 180))

# output numeric vector instead of dataframe
above_percent(example_data_1_subject)

data(example_data_5_subject)

above_percent(example_data_5_subject)
above_percent(example_data_5_subject, targets_above = c(70, 170))

}
\references{
Rodbard (2009) Interpretation of continuous glucose monitoring data:
glycemic variability and quality of glycemic control,
\emph{Diabetes Technology and Therapeutics} \strong{11 Suppl 1},
S55-67. \doi{10.1089/dia.2008.0132}.
}
