% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrr.R
\name{adrr}
\alias{adrr}
\title{Calculate average daily risk range (ADRR)}
\usage{
adrr(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}
}
\value{

}
\description{
Calculate average daily risk range (ADRR)
}
\examples{
adrr(data)

}
