% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrr.R
\name{adrr}
\alias{adrr}
\title{Calculate average daily risk range (ADRR)}
\usage{
adrr(data)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. Missing values will be linearly
interpolated when close enough to non-missing values.}
}
\value{
A data.frame of values of ADRR
}
\description{
The function adrr produces ADRR values in data.frame form
with one column and one row per subject.
}
\details{
A dataframe structure with one column and a row for each subject.

ADRR is calculated by \eqn{1/M * \sum [LR^i + HR^i]}
where M is number of days, LR is the max low risk value for day i
and HR is the max high risk value for day i.

Wrapping as.numeric() around the adrr call on a dataset with
a single subject will return a numeric value corresponding
to the ADRR value. This will not work for datasets with multiple subjects.
}
\examples{

data(example_data_1_subject)
adrr(example_data_1_subject)

data(example_data_5_subject)
adrr(example_data_5_subject)

}
\references{
Kovatchev et al. (2006) Evaluation of a New Measure of Blood Glucose Variability in,
Diabetes
\emph{Diabetes care} \strong{29} .2433-2438,
\doi{10.2337/dc06-1085}.
}
