% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_metrics.R
\name{all_metrics}
\alias{all_metrics}
\title{Calculate all metrics in iglu}
\usage{
all_metrics(data)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl".}
}
\value{
If a data.frame object is passed, then a tibble object with 1 row for each subject, and 49 columns is returned:
a column for subject id,
a column for adrr value,
a column for hourly_auc value,
a column for CONGA(24) value,
a column for cv value,
a column for CVmean value,
a column for CVsd value,
a column for ea1c value,
a column for gmi value,
a column for grade value,
a column for grade_eugly value,
a column for grade_hyper value,
a column for grade_hypo value,
a column for gvp value,
a column for hbgi value,
a column for lbgi value,
a column for hyper_index value,
a column for hypo_index value,
a column for igc value,
a column for iqr value,
a column for j_index value,
a column for m_value value,
a column for MAD value,
a column for mage value,
a column for modd value,
a column for above_140 value,
a column for above_180 value,
a column for above_200 value,
a column for above_250 value,
a column for below_54 value,
a column for below_70 value,
a column for in_range_70_140 value,
a column for in_range_70_180 value,
a column for in_range_80_200 value,
a column for range value,
a column for sd value,
a column for SdW value,
a column for SdHHMM value,
a column for SdWSH value,
a column for SdDM value,
a column for SdB value,
a column for SdBDM value,
a column for sd_roc value,
a column for Min. value,
a column for 1st Qu. value,
a column for Median value,
a column for Mean value,
a column for 3rd Qu. value,
a column for Max. value.
}
\description{
The function all_metrics runs all of the iglu functions:
ADRR, AUC, CONGA, CV_GLU, CV_Measures, eA1C, GMI, GRADE,
GRADE_Euglycemia, GRADE_Hyperglycemia, GRADE_Hypoglycemia,
GVP, HBGI, LBGI, Hyper_Index, Hypo_Index, IGC, IQR_GLU,
J_Index, M_Value, Mad_GLU, MAGE, MODD, Percent_Above,
Percent_Below, Percent_In_Range, Range, SD_GLU,
SD_Measures, SD_ROC, Summary
and combines them into a tibble object.
}
\details{
Returns a tibble object with 1 row for each subject, and 49 columns:
a column for subject id,
a column for adrr value,
a column for hourly_auc value,
a column for CONGA(24) value,
a column for cv value,
a column for CVmean value,
a column for CVsd value,
a column for ea1c value,
a column for gmi value,
a column for grade value,
a column for grade_eugly value,
a column for grade_hyper value,
a column for grade_hypo value,
a column for gvp value,
a column for hbgi value,
a column for lbgi value,
a column for hyper_index value,
a column for hypo_index value,
a column for igc value,
a column for iqr value,
a column for j_index value,
a column for m_value value,
a column for MAD value,
a column for mage value,
a column for modd value,
a column for above_140 value,
a column for above_180 value,
a column for above_200 value,
a column for above_250 value,
a column for below_54 value,
a column for below_70 value,
a column for in_range_70_140 value,
a column for in_range_70_180 value,
a column for in_range_80_200 value,
a column for range value,
a column for sd value,
a column for SdW value,
a column for SdHHMM value,
a column for SdWSH value,
a column for SdDM value,
a column for SdB value,
a column for SdBDM value,
a column for sd_roc value,
a column for Min. value,
a column for 1st Qu. value,
a column for Median value,
a column for Mean value,
a column for 3rd Qu. value,
a column for Max. value.
}
\examples{
data(example_data_1_subject)
all_metrics(example_data_1_subject)

}
