% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conga.R
\name{conga}
\alias{conga}
\title{Calculate continuous overall net glycemic action (CONGA)}
\usage{
conga(data)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. Missing values will be linearly
interpolated when close enough to non-missing values.}
}
\value{

}
\description{
The function conga produces a CONGA values in data.frame form
with one column and one row per subject. conga currently only supports
calculation of CONGA24.
}
\details{
A dataframe structure with one column and a row for each subject.

CONGA24 is currently the only supported CONGA type.

Wrapping as.numeric() around the above_percent call on a dataset with
a single subject will return a numeric vector, where the values
correspond to the percent of glucose values above each threshold
in the order passed in the targets argument. This will not work for
datasets with multiple subjects.
}
\examples{

data(example_data_1_subject)
conga(example_data_1_subject)

data(example_data_5_subject)
conga(example_data_5_subject)

}
\references{
Rodbard (2009) Interpretation of continuous glucose monitoring data:
glycemic variability and quality of glycemic control,
\emph{Diabetes Technology and Therapeutics} \strong{11 Suppl 1},
S55-67. \doi{10.1089/dia.2008.0132}.
}
