% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conga.R
\name{conga}
\alias{conga}
\title{Calculate continuous overall net glycemic action (CONGA)}
\usage{
conga(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}
}
\value{

}
\description{
Calculate continuous overall net glycemic action (CONGA)
}
\examples{
conga(data)
}
