% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_glu.R
\name{cv_glu}
\alias{cv_glu}
\title{Calculate Coefficient of Variation (CV) of glucose levels}
\usage{
cv_glu(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}
}
\value{

}
\description{
The coefficient of variation is computed by dividing
the sample standard deviation by the sample mean and multypling by 100%
}
\examples{
cv_glu(data)
}
