% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_profile.R
\name{epicalc_profile}
\alias{epicalc_profile}
\title{Display Episode Calculation statistics for selected subject}
\usage{
epicalc_profile(data, maxd = 14, inter_gap = 45, dt0 = NULL, tz = "", daily = TRUE)
}
\value{
A plot displaying the varying glucose levels (mg/dL) of the subject in a day as well as the statistics for the episodes.
}
\description{
Display Episode Calculation statistics for selected subject
}
\examples{
epicalc_profile(example_data_1_subject)

}
\author{
Johnathan Shih, Jung Hoon Seo
}
