% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_profile.R
\name{epicalc_profile}
\alias{epicalc_profile}
\title{Display Episode Calculation statistics for selected subject}
\usage{
epicalc_profile(data, lv1_hypo=100.0, lv2_hypo = 70, lv1_hyper= 120.0, lv2_hyper = 160, color_scheme = "Color Scheme 1")
}
\arguments{
\item{data}{data}

\item{lv1_hypo}{hypo_thres}

\item{lv2_hypo}{hypo_thres}

\item{lv1_hyper}{lv1_hyper}

\item{lv2_hyper}{lv2_hyper}

\item{color_scheme}{color_scheme}
}
\value{
A plot displaying the varying glucose levels (mg/dL) of the subject in a day as well as the statistics for the episodes.
}
\description{
Display Episode Calculation statistics for selected subject
}
\examples{
epicalc_profile(example_data_1_subject)

}
\author{
Johnathan Shih, Jung Hoon Seo
}
