% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_final.R
\name{episode_calculation}
\alias{episode_calculation}
\title{Calculates the number of Hypo/Hyperglycemic events as well as other statistics}
\usage{
episode_calculation(data, hypo_thres=90.0, hyper_thres= 120.0, dur_length = 15)
}
\value{
Data frame including Average Glucose, number of hypo and hyper episodes, Hypo and hyper mean values, the percentages of low, high and target alerts
}
\description{
The function episode calculation produces the number of Hypo/Hyperglycemic events as well as other statistics such as the average glucose level, mean duration, and the percentage of time spent in the ranges.
}
\details{

}
\examples{

episode_calculation(example_data_4_subject, dt0 = 5)

}
\references{

}
\author{
Johnathan Shih, Jung Hoon Seo
}
