% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_meals_hall}
\alias{example_meals_hall}
\title{Example mealtimes data from Hall et al. (2018)}
\format{
A data.frame with 9 rows and 3 columns, which are:
\describe{
\item{id}{identifier of subject}
\item{meal}{meal type identifier}
\item{mealtime}{time of meal}
}
}
\usage{
example_meals_hall
}
\description{
Example of mealtimes data format for meal_metrics function,
corresponds to example_data_hall data.
}
\references{
Hall et al. (2018) : Glucotypes reveal new patterns of glucose dysregulation
\emph{Plos Biology} \strong{16} (7): 3:e2005143
\doi{10.1371/journal.pbio.2005143}.
}
\keyword{datasets}
