% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{grade}
\alias{grade}
\title{Calculate mean GRADE score}
\usage{
grade(data)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. NA's will be omitted from the glucose
values in calculation of GRADE score.}
}
\value{

}
\description{
The function grade produces GRADE score values in data.frame form
with one column and one row per subject.
}
\details{
A dataframe structure with one column and a row for each subject.

GRADE score is calculated by \eqn{1/n * \sum [425 *
(log(log(BG_i / 18)) + .16)^2]}
Where \eqn{BG_i} is the ith Blood Glucose measurement and n is the total
number of measurements.

Wrapping as.numeric() around the grade call on a dataset with
a single subject will return a numeric value corresponding to the GRADE
score. This will not work for datasets with multiple subjects.
}
\examples{

data(example_data_1_subject)
grade(example_data_1_subject)

data(example_data_5_subject)
grade(example_data_5_subject)

}
\references{
Hill et al. (2007): A method for assessing quality of control
from glucose profiles
\emph{Diabetic Medicine} \strong{24} .753-758,
\doi{10.1111/j.1464-5491.2007.02119.x}.
}
