% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_eugly.R
\name{grade_eugly}
\alias{grade_eugly}
\title{Percentage of GRADE score attributable to target range}
\usage{
grade_eugly(data, lower = 70, upper = 140)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. NA's will be omitted from the glucose
values in calculation of \%GRADE euglycemia.}

\item{lower}{Lower bound used for hypoglycemia cutoff, in mg/dL. Default is 70}

\item{upper}{Upper bound used for hyperglycemia cutoff, in mg/dL. Default is 140.}
}
\value{

}
\description{
The function grade_eugly produces \%GRADE euglycemia values in data.frame form
with one column and one row per subject.
}
\details{
A dataframe structure with one column and a row for each subject.

\%GRADE euglycemia is determined by calculating the percentage of GRADE score (see grade
function) attributed to values in the target range, i.e. values not below
hypoglycemic or above hyperglycemic cutoffs.

Wrapping as.numeric() around the grade_eugly call on a dataset with
a single subject will return a numeric value corresponding to the \%GRADE
euglycemia value. This will not work for datasets with multiple subjects.
}
\examples{

data(example_data_1_subject)
grade_eugly(example_data_1_subject)
grade_eugly(example_data_1_subject, lower = 80, upper = 180)

data(example_data_5_subject)
grade_eugly(example_data_5_subject)
grade_eugly(example_data_5_subject, lower = 80, upper = 160)

}
\references{
Rodbard (2009) Interpretation of continuous glucose monitoring data:
glycemic variability and quality of glycemic control,
\emph{Diabetes Technology and Therapeutics} \strong{11 Suppl 1},
S55-67. \doi{10.1089/dia.2008.0132}.
}
