% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_hyper.R
\name{grade_hyper}
\alias{grade_hyper}
\title{Percentage of GRADE score above target}
\usage{
grade_hyper(data, upper = 140)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{upper}{Upper bound used for hyperglycemia cutoff, in mg/dL. Default is 140.}
}
\value{

}
\description{
Percentage of GRADE score above target
}
\examples{
grade_hyper(data)
grade_hyper(data, upper = 160)

}
