% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_eugly.R, R/grade_hyper.R, R/grade_hypo.R
\name{grade_hyper}
\alias{grade_hyper}
\title{Percentage of GRADE score in target range}
\usage{
grade_hyper(data, lower = 80)

grade_hyper(data, lower = 80)

grade_hyper(data, lower = 80)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{lower}{Lower bound used for hypoglycemia cutoff, in mg/dL. Default is 80}

\item{upper}{Upper bound used for hyperglycemia cutoff, in mg/dL. Default is 140.}

\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{upper}{Upper bound used for hyperglycemia cutoff, in mg/dL. Default is 140.}

\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{lower}{Lower bound used for hypoglycemia cutoff, in mg/dL. Default is 80}
}
\value{





}
\description{
Percentage of GRADE score in target range

Percentage of GRADE score above target

Percentage of GRADE score below target
}
\examples{
grade_hyper(data)
grade_hyper(data, lower = 70)
grade_hyper(data)
grade_hyper(data, upper = 160)
grade_hypeodata)
grade_hypo(data, lower = 70)
}
