% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_hypo.R
\name{grade_hypo}
\alias{grade_hypo}
\title{Percentage of GRADE score below target}
\usage{
grade_hypo(data, lower = 70)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{lower}{Lower bound used for hypoglycemia cutoff, in mg/dL. Default is 80}
}
\value{

}
\description{
Percentage of GRADE score below target
}
\examples{
grade_hypo(data)
grade_hypo(data, lower = 80)
}
