% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbgi.R
\name{hbgi}
\alias{hbgi}
\title{Calculate High Blood Glucose Index (HGBI)}
\usage{
hbgi(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}
}
\description{
Calculate High Blood Glucose Index (HGBI)
}
\examples{
hbgi(data)
}
