% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_index.R
\name{hyper_index}
\alias{hyper_index}
\title{Calculate hyperglycaemic index (HGI)}
\usage{
hyper_index(data, upper = 140)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}

\item{upper}{Upper bound used for hyperglycemia cutoff. Default is 140.}
}
\value{

}
\description{
Calculate hyperglycaemic index (HGI)
}
\examples{
hyper_index(data)
hyper_index(data, upper = 160)
}
