% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypo_index.R
\name{hypo_index}
\alias{hypo_index}
\title{Calculate hypoglycaemic index (HGI)}
\usage{
hypo_index(data, lower = 80)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}

\item{lower}{Lower bound used for hypoglycemia cutoff. Default is 80.}
}
\value{

}
\description{
Calculate hypoglycaemic index (HGI)
}
\examples{
hypo_index(data)
hypo_index(data, upper = 70)
}
