% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypo_index.R
\name{hypo_index}
\alias{hypo_index}
\title{Calculate Hypoglycemia Index}
\usage{
hypo_index(data, LLTR = 70)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values.}

\item{LLTR}{Lower Limit of Target Range, default value is 80 mg/dL.}

\item{b}{Exponent, generally in the range from 1.0 to 2.0, default value is 2.}

\item{d}{Scaling factor,to display Hyperglycemia Index, Hypoglycemia Index, and IGC on approximately the same numerical range as measurements of HBGI, LBGI and GRADE, default value is 30.}
}
\value{
If a data.frame object is passed, then a tibble object with
two columns: subject id and corresponding Hypoglycemia Index value is returned. If a vector of glucose
values is passed, then a tibble object with just the Hypoglycemia Index value is returned.
as.numeric() can be wrapped around the latter to ouput just a numeric value.
}
\description{
The function hypo_index produces Hypoglycemia index values in a tibble object.
}
\details{
A tibble object with 1 row for each subject, a column for subject id and
a column for the Hypoglycemia Index values is returned. NA glucose values are
omitted from the calculation of the Hypoglycemia Index values.

Hypoglycemia Index is calculated by \eqn{n/d * \sum [(LLTR-hypoBG_j)^{b}]}
Here n is the total number of Blood Glucose measurements (excluding NA values),
and \eqn{hypoBG_j} is the jth Blood Glucose measurement below the LLTR cutoff, b is an exponent, and d is a scaling factor.
}
\examples{
data(example_data_1_subject)
hypo_index(example_data_1_subject, LLTR = 60)

data(example_data_5_subject)
hypo_index(example_data_5_subject)
hypo_index(example_data_5_subject, LLTR = 70)

}
\references{
Rodbard (2009) Interpretation of continuous glucose monitoring data:
glycemic variability and quality of glycemic control,
\emph{Diabetes Technology and Therapeutics} \strong{11} .55-67,
\doi{10.1089/dia.2008.0132}.
}
