% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igc.R
\name{igc}
\alias{igc}
\title{Compute sum of Hyperglycemia index and Hypoglycemia index}
\usage{
igc(data, lower = 80, upper = 140)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{lower}{Lower bound used for hypoglycemia cutoff. Default is 80.}

\item{upper}{Upper bound used for hyperglycemia cutoff. Default is 140.}
}
\value{

}
\description{
Compute sum of Hyperglycemia index and Hypoglycemia index
}
\examples{
igc(data)
igc(data, lower = 70, upper = 160)
}
