% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_range_percent.R
\name{in_range_percent}
\alias{in_range_percent}
\title{Calculate percentage in targeted value ranges}
\usage{
in_range_percent(data, targets = list(c(80, 200), c(70, 180), c(70,
  140)))
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}

\item{targets}{List of target values. Default list is [(80,200), (70,180), (70,140)].}
}
\value{

}
\description{
Calculate percentage in targeted value ranges
}
\examples{
in_range_percent(data)
in_range_percent(data, targets = list(c(100,200), c(150,250))

# The following will return same output as previous output:
in_range_percent(data, targets = list(c(200, 100)), c(250,150))
}
