% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_glu.R
\name{iqr_glu}
\alias{iqr_glu}
\title{Calculate glucose level iqr}
\usage{
iqr_glu(data)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. NA's will be omitted from the glucose
values in calculation of the iqr.}
}
\value{

}
\description{
The function iqr_glu outputs the distance between the 25th
percentile and the 25th percentile of the glucose values for each subject.
The output is in a data.frame form by default, with one column and a
row corresponding to each subject.
}
\details{
A dataframe structure with 1 row for each subject and 1 column
for the iqr value is returned.

Wrapping as.numeric() around the iqr_glu call on a dataset with
a single subject will return a numeric value corresponding to the iqr.
This will not work for datasets with multiple subjects.
}
\examples{
data(example_data_1_subject)
iqr_glu(example_data_1_subject)

data(example_data_5_subject)
iqr_glu(example_data_5_subject)

}
