% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_index.R
\name{j_index}
\alias{j_index}
\title{Calculate J-index}
\usage{
j_index(data)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. NA's will be omitted from the glucose
values in calculation of J-index score.}
}
\value{

}
\description{
The function j_index produces J-index values in data.frame form
with one column and one row per subject.
}
\details{
A dataframe structure with one column and a row for each subject.

J-index score is calculated by \eqn{.001 * [mean(BG) + sd(BG)]^2}
where BG is the list of Blood Glucose Measurements.

Wrapping as.numeric() around the j_index call on a dataset with
a single subject will return a numeric value corresponding to the J-index
score. This will not work for datasets with multiple subjects.
}
\examples{

data(example_data_1_subject)
j_index(example_data_1_subject)

data(example_data_5_subject)
j_index(example_data_5_subject)

}
\references{
Wojcicki (1995) "J"-index. A new proposition of the assessment
of current glucose control in diabetic patients
\emph{Hormone and Metabolic Research} \strong{27} .41-42,
\doi{10.1055/s-2007-979906}.
}
