% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbgi.R
\name{lbgi}
\alias{lbgi}
\title{Calculate Low Blood Glucose Index (LBGI)}
\usage{
lbgi(data)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. NA's will be omitted from the glucose
values in calculation of LBGI.}
}
\value{
A data.frame with the index/measure and identifier column, if
the input data is a `data.frame`
}
\description{
The function lbgi produces LBGI values in data.frame form
with one column and one row per subject.
}
\details{
A dataframe structure with one column and a row for each subject.

LBGI is calculated by \eqn{1/n * \sum (10 * fbg_i ^2)},
where \eqn{fbg_i = min(0, 1.509 * (log(BG_i)^{1.084} - 5.381)},
BG_i is the ith Blood Glucose measurement for a subject, and
n is the total number of measurements for that subject.

Wrapping as.numeric() around the lbgi call on a dataset with
a single subject will return a numeric value corresponding
to the LBGI value. This will not work for datasets with multiple subjects.
}
\examples{

data(example_data_1_subject)
lbgi(example_data_1_subject)

data(example_data_5_subject)
lbgi(example_data_5_subject)

}
\references{
Kovatchev et al. (2006) Evaluation of a New Measure of Blood Glucose Variability in,
Diabetes
\emph{Diabetes care} \strong{29} .2433-2438,
\doi{10.2337/dc06-1085}.
}
