% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mage.R
\name{mage}
\alias{mage}
\title{Calculate MAGE (Mean amplitude of glucose excursion)}
\usage{
mage(data, id = unique(data$id), sd = 1)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}

\item{id}{Desired value(s) from id column. Default is all id's.}

\item{sd}{Standard deviations used to classify "excursions". Default is 1.}
}
\description{
Calculate MAGE (Mean amplitude of glucose excursion)
}
\examples{
mage(mydata)
mage(mydata, 'Subject 1', 2)
}
