% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mage.R
\name{mage}
\alias{mage}
\title{Calculate Mean Amplitude of Glucose Excursion (MAGE)}
\usage{
mage(data, sd = 1)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}

\item{sd}{Standard deviations used to classify "excursions". Default is 1.}
}
\value{

}
\description{
Calculate Mean Amplitude of Glucose Excursion (MAGE)
}
\examples{
mage(data)
mage(data, 'Subject 1', 2)

}
