% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_scatterplot.R
\name{metric_scatter}
\alias{metric_scatter}
\title{Plot scatterplot of metric comparisons}
\usage{
metric_scatter(
  data,
  metric1 = c("adrr", "grade_hypo", "hypo_index", "grade_hyper", "hyper_index",
    "grade_eugly", "igc", "mage", "modd"),
  metric2 = "conga",
  subjects = NULL,
  lower = 70,
  upper = 140,
  targets_above = c(140, 180, 200, 250),
  targets_below = c(50, 80),
  target_ranges = list(c(80, 200), c(70, 180), c(70, 140)),
  quantiles = c(0, 25, 50, 75, 100),
  sd_multiplier = 1,
  lag = 1
)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}

\item{metric2}{}
}
\value{

}
\description{
Plot scatterplot of metric comparisons
}
