% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{metrics_heatmap}
\alias{metrics_heatmap}
\title{Create a heatmap that clusters metrics and subjects according to the metrics}
\usage{
metrics_heatmap(
  mecs,
  metric_cluster = 6,
  clustering_method = "complete",
  clustering_distance_metrics = "correlation",
  clustering_distance_subjects = "correlation"
)
}
\arguments{
\item{mecs}{precalculated metric values, with first column corresponding to subject id}

\item{metric_cluster}{number of visual metric clusters, default value is 6}

\item{clustering_method}{the agglomeration method for hierarchica clustering, accepts same values as hclust, default value is "complete"}

\item{clustering_distance_metrics}{the distance measure for metrics clustering, accepts same values as dist, default value is "correlation" distance}

\item{clustering_distance_subjects}{the distance measure for subjects clustering, accepts same values as dist, default value is "correlation" distance}
}
\value{
A heatmap of metrics by subjects generated by pheatmap
}
\description{
Create a heatmap that clusters metrics and subjects according to the metrics
}
\examples{
mecs = all_metrics(example_data_5_subject)
p = metrics_heatmap(mecs)
print(p)
}
