% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modd.R
\name{modd}
\alias{modd}
\title{Calculate mean difference between glucose values obtained at the same time
of day (MODD)}
\usage{
modd(data, lag = 1)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl",
or numeric vector of glucose values. Missing values will be linearly
interpolated when close enough to non-missing values.}

\item{lag}{Integer indicating which lag (# days) to use. Default is 1.}
}
\value{

}
\description{
The function modd produces MODD values in data.frame form
with one column and one row per subject.
}
\details{
A dataframe structure with one column and a row for each subject.

MODD is calculated by taking the mean of absolute differences between
measurements at the same time 1 day away, or more if lag parameter
is set to an integer > 1.

Wrapping as.numeric() around the modd call on a dataset with
a single subject will return a numeric value corresponding
to the MODD value. This will not work for datasets with multiple subjects.
}
\examples{

data(example_data_1_subject)
modd(example_data_1_subject)
modd(example_data_1_subject, lag = 2)

data(example_data_5_subject)
modd(example_data_5_subject, lag = 2)

}
\references{
Service, Nelson (1980) Characteristics of glycemic stability.
\emph{Diabetes care} \strong{3} .58-62,
\doi{10.2337/diacare.3.1.58}.
}
