% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glu.R
\name{plot_glu}
\alias{plot_glu}
\title{Plot time series and lasagna plots of glucose measurements}
\usage{
plot_glu(
  data,
  plottype = c("tsplot", "lasagna"),
  datatype = c("all", "average", "single"),
  lasagnatype = c("unsorted", "timesorted"),
  LLTR = 80,
  ULTR = 140,
  subjects = NULL,
  tz = ""
)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}

\item{plottype}{String corresponding to the desired plot type. Options are
'tsplot' for a time series plot and 'lasagna' for a lasagna plot. See the
lasagnatype parameter for further options corresponding to the 'lasagna' plottype.}

\item{datatype}{String corresponding to data aggregation used for plotting, currently supported
options are 'all' which plots all glucose measurements within the first \code{maxd} days for each subject, and 'average' which plots average 24 hour glucose values across days for each subject}

\item{lasagnatype}{String corresponding to plot type when using\code{datatype = "average"}, currently supported options are 'unsorted' for an unsorted lasagna plot, and 'timesorted' for a
lasgna plot with glucose values sorted for each time point across subjects}

\item{LLTR}{Lower Limit of Target Range, default value is 80 mg/dL.}

\item{ULTR}{Upper Limit of Target Range, default value is 140 mg/dL.}

\item{subjects}{String or list of strings corresponding to subject names
in 'id' column of data.}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link[base]{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.}
}
\value{
Any output from the plot object
}
\description{
The function plot_glu contains several plotting types for single and multiple
subject data.
}
\details{
For the 'tsplot' option, only the first subject will be used, whether the subjects
argument is left NULL or a subject list is provided. To choose a particular subject,
it is best to only list that subject's id in the subjects argument of plot_glu.
The lower and upper values are shown as horizontal lines on the plot to show
the target range.

For the 'unsorted' and 'rowsorted' options, all subjects are used if the subjects
argument is left NULL. If a list of subject IDs is provided, all of those will
be used.
}
\examples{

data(example_data_1_subject)
plot_glu(example_data_1_subject)
plot_glu(example_data_1_subject, LLTR = 70)

data(example_data_5_subject)
plot_glu(example_data_5_subject, subjects = 'Subject 2')
plot_glu(example_data_5_subject, plottype = 'tsplot', tz = 'EST', LLTR = 70, ULTR= 150)
plot_glu(example_data_5_subject, plottype = 'lasagna', lasagnatype = 'timesorted')
plot_glu(example_data_5_subject, plottype = 'lasagna', lasagnatype = 'unsorted')

}
