% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glu.R
\name{plot_glu}
\alias{plot_glu}
\title{Plot time series of glucose measurements}
\usage{
plot_glu(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}
}
\value{

}
\description{
Plot time series of glucose measurements
}
\examples{
plot_glu(data)

}
