% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glu.R
\name{plot_glu}
\alias{plot_glu}
\title{Plot time series and lasagna plots of glucose measurements}
\usage{
plot_glu(
  data,
  plottype = c("tsplot", "unsorted", "rowsorted"),
  lower = 70,
  upper = 140,
  subjects = NULL
)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}

\item{plottype}{String corresponding to plot type, currently supported
options are 'tsplot' for a 1 subject time series line plot, 'unsorted'
for an unsorted multi-subject lasagna plot, and 'rowsorted' for a
row-sorted lasagna plot.}

\item{lower}{Numeric value for hypoglycemia cutoff}

\item{upper}{Numeric value for hyperglycemia cutoff}

\item{subjects}{String or list of strings corresponding to subject names
in 'id' column of data.}
}
\value{
Any output from the plot object
}
\description{
The function plot_glu contains several plotting types for single and multiple
subject data.
}
\details{
For the 'tsplot' option, only the first subject will be used, whether the subjects
argument is left NULL or a subject list is provided. To choose a particular subject,
it is best to only list that subject's id in the subjects argument of plot_glu.
The lower and upper values are shown as horizontal lines on the plot to show
the target range.

For the 'unsorted' and 'rowsorted' options, all subjects are used if the subjects
argument is left NULL. If a list of subject IDs is provided, all of those will
be used.
}
\examples{

data(example_data_1_subject)
plot_glu(example_data_1_subject)
plot_glu(example_data_1_subject, lower = 80)

data(example_data_5_subject)
plot_glu(example_data_5_subject, subjects = 'Subject 2')
plot_glu(example_data_5_subject, plottype = 'unsorted')
plot_glu(example_data_5_subject, plottype = 'rowsorted')

}
