% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot time series of glucose colored by rate of change}
\usage{
plot_roc(data, subjects = NULL, timelag = 15, tz = "")
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", and "gl".}

\item{subjects}{String or list of strings corresponding to subject names
in 'id' column of data. Default is all subjects.}

\item{timelag}{Integer indicating the time period (# minutes) over which rate
of change is calculated. Default is 15, e.g. rate of change is the change in
glucose over the past 15 minutes divided by 15.}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.}
}
\value{
A time series of glucose values colored by ROC categories per subject
}
\description{
The function plot_roc produces a time series plot of glucose values colored
by categorized rate of change values
}
\details{
For the default, a time series plot is produced for each subject in which the
glucose values are plotted and colored by the following categories for ROC:
breaks = c(-Inf, -3, -2, -1, 1, 2, 3, Inf) where the glucose is in mg/dl and
the ROC values are in mg/dl/min. A ROC of -5 mg/dl/min will thus be placed in
category 1 and colored accordingly.
}
\examples{

data(example_data_1_subject)
plot_roc(example_data_1_subject)

data(example_data_5_subject)
plot_roc(example_data_5_subject)
plot_roc(example_data_5_subject, subjects = 'Subject 5')

}
