% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_glu.R
\name{quantile_glu}
\alias{quantile_glu}
\title{Calculate glucose level quantiles}
\usage{
quantile_glu(data, quantiles = c(0, 0.25, 0.5, 0.75, 1))
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}

\item{quantiles}{List of values between 0 and 1.}
}
\description{
Simple wrapper for quantile().
}
\examples{
quantile_glu(data)
quantile_glu(data, quantiles = c(0, .33, .66, 1))
}
