% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_glu.R
\name{range_glu}
\alias{range_glu}
\title{Calculate glucose level range}
\usage{
range_glu(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}
}
\value{

}
\description{
Computes distance between min and max glucose value.
}
\examples{
range_glu(data)
}
