% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_glu.R
\name{sd_glu}
\alias{sd_glu}
\title{Calculate standard deviation of glucose levels}
\usage{
sd_glu(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl").}
}
\value{

}
\description{
Simple wrapper for sd().
}
\examples{
sd_glu(data)
}
