% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_glu.R
\name{summary_glu}
\alias{summary_glu}
\title{Calculate summary statistics of glucose levels}
\usage{
summary_glu(data)
}
\arguments{
\item{data}{DataFrame with column names ("id", "time", and "gl"),
or vector of glucose values as integer, numeric, or double.}
}
\value{

}
\description{
Simple wrapper for summary().
}
\examples{
summary_glu(data)
}
